refactor: refactor frontend structure
