--- frontend/src/app/admin/students/[studentId]/(studentInfo)/page.tsx
@@ -1,5 +1,0 @@
-import StudentInfo from "@/components/pages/StudentInfo";
-
-export default function StudentInfoPage() {
-  return <StudentInfo />;
-}
