+++ frontend/src/app/admin/students/[studentId]/page.tsx
@@ -1,0 +1,5 @@
+import StudentInfo from "@/features/students/components/profile/StudentInfo";
+
+export default function StudentInfoPage() {
+  return <StudentInfo />;
+}
