+++ frontend/src/app/admin/students/page.tsx
@@ -1,4 +1,4 @@
-import StudentTable from "@/components/tables/StudentTable";
+import StudentTable from "@/features/students/components/table/StudentTable";
 
 export default function StudentDatabase() {
   return <StudentTable />;
