+++ frontend/src/features/attendances/services/attendanceService.ts
@@ -1,0 +1,93 @@
+import { api } from "@/services";
+import { attendanceMock } from "@/services/mocks";
+import {
+  Attendance,
+  AttendanceFormData,
+  PaginatedAttendancesResponse,
+} from "../types/attendance";
+
+export const attendanceService = {
+  async getAttendances(
+    page: number = 1,
+    limit: number = 10,
+  ): Promise<PaginatedAttendancesResponse> {
+    const response = await api.get<PaginatedAttendancesResponse>(
+      "/attendances",
+      {
+        params: { page, limit },
+      },
+    );
+
+    return response.data;
+  },
+
+  async getAttendancesById(id: string): Promise<Attendance> {
+    try {
+      const response = await api.get<Attendance>(`/attendances/${id}`);
+      return response.data;
+    } catch {
+      const attendance = attendanceMock.find((s) => s.attendanceId === id);
+
+      if (!attendance) {
+        throw new Error(`Attendance with ID ${id} not found in mock.`);
+      }
+
+      return attendance;
+    }
+  },
+
+  async getAttendancesByStudent(
+    studentId: string,
+    page: number = 1,
+    limit: number = 10,
+  ): Promise<PaginatedAttendancesResponse> {
+    const response = await api.get<PaginatedAttendancesResponse>(
+      `/attendances/student/${studentId}`,
+      {
+        params: { page, limit },
+      },
+    );
+
+    return response.data;
+  },
+
+  async getAttendancesByStudentt(
+    studentId: string,
+    page: number,
+    limit: number,
+  ) {
+    try {
+      const response = await api.get(`/attendances/student/${studentId}`, {
+        params: { page, limit },
+      });
+      return response.data;
+    } catch {
+      const attendance = attendanceMock.filter(
+        (s) => s.studentId === studentId,
+      );
+
+      if (!attendance) {
+        throw new Error(`Attendance with ID ${studentId} not found in mock.`);
+      }
+
+      return attendance;
+    }
+  },
+
+  async create(data: AttendanceFormData): Promise<AttendanceFormData> {
+    const response = await api.post<AttendanceFormData>("/attendances", data);
+    return response.data;
+  },
+
+  async update(
+    id: string,
+    data: Partial<AttendanceFormData>,
+  ): Promise<Attendance> {
+    const response = await api.put<Attendance>(`/attendances/${id}`, data);
+    return response.data;
+  },
+
+  async removeAttendance(attendanceId: string): Promise<void> {
+    await api.post(`/attendances/${attendanceId}/remove`);
+  },
+};
