+++ frontend/src/services/index.ts
@@ -1,2 +1,1 @@
 export { default as api } from "./api";
-export { studentService } from "./studentService";
