--- frontend/src/types/specialNeed.ts
@@ -1,6 +1,0 @@
-export type SpecialNeed =
-  | "TDAH TAG"
-  | "TEA"
-  | "PCD"
-  | "Dificuldade de aprendizado"
-  | "Nenhuma";
